внутреннее(контроллер) добавлен контроллер для авторизации через вк

+++ src/main/java/app/repository/ActionRepository.java
@@ -1,7 +1,7 @@
 package app.repository;
 
-import app.model.Action;
+import app.entity.Action;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionRepository extends CrudRepository<Action,Integer> {
+public interface ActionRepository extends CrudRepository<Action,Long> {
 }
