+++ src/main/java/app/repository/CustomerRepository.java
@@ -1,7 +1,7 @@
 package app.repository;
 
-import app.model.Customer;
+import app.entity.Customer;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customer,Integer> {
+public interface CustomerRepository extends CrudRepository<Customer,Long> {
 }
